# golang-gin/examples/simpleshop/internal/order/model/user.go
package model

import (
	"time"
)

type User struct {
	ID        string     `json:"id" gorm:"unique;not null;index;primary_key"`
	Email     string     `json:"email" gorm:"unique;not null;index:idx_user_email"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt time.Time  `json:"updated_at"`
	DeletedAt *time.Time `json:"deleted_at" gorm:"index"`
}